Split rule processing logic out of RuleRegistry
